final refactor and starting with SuiClient
