+++ packages/reown_walletkit/lib/i_walletkit_impl.dart
@@ -1,6 +1,7 @@
 import 'package:reown_sign/i_sign_wallet.dart';
 import 'package:reown_walletkit/chain_abstraction/i_chain_abstraction.dart';
 import 'package:reown_walletkit/reown_walletkit.dart';
+import 'package:reown_walletkit/sui/i_sui_client.dart';
 
 abstract class IReownWalletKit
     implements IReownSignWallet, IChainAbstractionClient {
@@ -8,5 +9,6 @@ abstract class IReownWalletKit
   final int version = 2;
 
   abstract final IReownSign reOwnSign;
-  abstract final IChainAbstractionClient reOwnChainAbstraction;
+  abstract final IChainAbstractionClient chainAbstractionClient;
+  abstract final ISuiClient suiClient;
 }
