+++ packages/reown_walletkit/lib/sui/i_sui_client.dart
@@ -1,0 +1,5 @@
+// import 'package:reown_walletkit/reown_walletkit.dart';
+
+abstract class ISuiClient {
+  Future<void> init();
+}
