+++ packages/reown_walletkit/lib/walletkit_impl.dart
@@ -7,6 +7,8 @@ import 'package:reown_walletkit/chain_abstraction/chain_abstraction.dart';
 import 'package:reown_walletkit/chain_abstraction/i_chain_abstraction.dart';
 
 import 'package:reown_walletkit/reown_walletkit.dart';
+import 'package:reown_walletkit/sui/i_sui_client.dart';
+import 'package:reown_walletkit/sui/sui_client.dart';
 import 'package:reown_walletkit/version.dart' as wk;
 
 class ReownWalletKit with WidgetsBindingObserver implements IReownWalletKit {
@@ -108,7 +110,16 @@ class ReownWalletKit with WidgetsBindingObserver implements IReownWalletKit {
       ),
     );
 
-    reOwnChainAbstraction = ChainAbstractionClient(
+    chainAbstractionClient = ChainAbstractionClient(
+      core: core,
+      pulseMetadata: PulseMetadataCompat(
+        url: metadata.url,
+        sdkVersion: wk.packageVersion,
+        sdkPlatform: ReownCoreUtils.getId(),
+      ),
+    );
+
+    suiClient = SuiClient(
       core: core,
       pulseMetadata: PulseMetadataCompat(
         url: metadata.url,
@@ -126,7 +137,8 @@ class ReownWalletKit with WidgetsBindingObserver implements IReownWalletKit {
 
     await core.start();
     await reOwnSign.init();
-    await reOwnChainAbstraction.init();
+    await chainAbstractionClient.init();
+    await suiClient.init();
 
     WidgetsBinding.instance.addObserver(this);
     _initialized = true;
@@ -509,7 +521,7 @@ class ReownWalletKit with WidgetsBindingObserver implements IReownWalletKit {
   ///---------- CHAIN ABSTRACTION CLIENT ----------///
   ///
   @override
-  late final IChainAbstractionClient reOwnChainAbstraction;
+  late final IChainAbstractionClient chainAbstractionClient;
 
   @override
   Future<String> erc20TokenBalance({
@@ -517,7 +529,7 @@ class ReownWalletKit with WidgetsBindingObserver implements IReownWalletKit {
     required String token,
     required String owner,
   }) async {
-    return await reOwnChainAbstraction.erc20TokenBalance(
+    return await chainAbstractionClient.erc20TokenBalance(
       chainId: chainId,
       token: token,
       owner: owner,
@@ -528,7 +540,7 @@ class ReownWalletKit with WidgetsBindingObserver implements IReownWalletKit {
   Future<Eip1559EstimationCompat> estimateFees({
     required String chainId,
   }) async {
-    return await reOwnChainAbstraction.estimateFees(
+    return await chainAbstractionClient.estimateFees(
       chainId: chainId,
     );
   }
@@ -545,7 +557,7 @@ class ReownWalletKit with WidgetsBindingObserver implements IReownWalletKit {
     Currency localCurrency = Currency.usd,
     bool useLifi = false,
   }) async {
-    return await reOwnChainAbstraction.prepare(
+    return await chainAbstractionClient.prepare(
       chainId: chainId,
       from: from,
       call: call,
@@ -564,10 +576,15 @@ class ReownWalletKit with WidgetsBindingObserver implements IReownWalletKit {
     required List<String> routeTxnSigs,
     required String initialTxnSig,
   }) async {
-    return await reOwnChainAbstraction.execute(
+    return await chainAbstractionClient.execute(
       uiFields: uiFields,
       routeTxnSigs: routeTxnSigs,
       initialTxnSig: initialTxnSig,
     );
   }
+
+  ///---------- SUI CLIENT ----------///
+  ///
+  @override
+  late final ISuiClient suiClient;
 }
