+++ packages/reown_yttrium/lib/clients/i_sui_client.dart
@@ -1,0 +1,8 @@
+import 'package:reown_yttrium/models/chain_abstraction.dart';
+
+abstract class ISuiClient {
+  Future<bool> init({
+    required String projectId,
+    required PulseMetadataCompat pulseMetadata,
+  });
+}
