+++ packages/reown_yttrium/lib/reown_yttrium_method_channel.dart
@@ -1,10 +1,14 @@
-import 'package:reown_yttrium/channels/chain_abstraction.dart';
+import 'package:reown_yttrium/channels/chain_abstraction_channel.dart';
+import 'package:reown_yttrium/channels/sui_channel.dart';
 
 import 'reown_yttrium_platform_interface.dart';
 
 /// An implementation of [ReownYttriumPlatform] that uses method channels.
 class MethodChannelReownYttrium extends ReownYttriumPlatformInterface {
   @override
-  MethodChannelChainAbstraction get chainAbstractionClient =>
+  MethodChannelChainAbstraction get chainAbstractionChannel =>
       MethodChannelChainAbstraction();
+
+  @override
+  MethodChannelSui get suiChannel => MethodChannelSui();
 }
